Reorganize and rename files
