+++ packages/twenty-sdk/src/cli/__tests__/e2e/constants/testConfig.ts
@@ -1,4 +1,4 @@
-import { type TwentyConfig } from '@/cli/utilities/config/services/config.service';
+import { type TwentyConfig } from '@/cli/utilities/config/config-service';
 
 export const testConfig: TwentyConfig = {
   apiUrl: 'http://localhost:3000',
