+++ packages/twenty-sdk/src/cli/commands/auth/auth-status.ts
@@ -1,6 +1,6 @@
 import chalk from 'chalk';
-import { ApiService } from '@/cli/utilities/api/services/api.service';
-import { ConfigService } from '@/cli/utilities/config/services/config.service';
+import { ApiService } from '@/cli/utilities/api/api-service';
+import { ConfigService } from '@/cli/utilities/config/config-service';
 
 export class AuthStatusCommand {
   private configService = new ConfigService();
