+++ packages/twenty-sdk/src/cli/commands/auth/auth-switch.ts
@@ -1,7 +1,7 @@
 import chalk from 'chalk';
 import inquirer from 'inquirer';
-import { ApiService } from '@/cli/utilities/api/services/api.service';
-import { ConfigService } from '@/cli/utilities/config/services/config.service';
+import { ApiService } from '@/cli/utilities/api/api-service';
+import { ConfigService } from '@/cli/utilities/config/config-service';
 
 export class AuthSwitchCommand {
   private configService = new ConfigService();
