--- packages/twenty-sdk/src/cli/constants/assets-dir.ts
@@ -1,5 +1,0 @@
-/**
- * Directory name for static assets in Twenty applications.
- * Assets are copied from this folder at the root of the app to the build output.
- */
-export const ASSETS_DIR = 'assets';
