+++ packages/twenty-sdk/src/cli/utilities/api/api-service.ts
@@ -1,4 +1,4 @@
-import { ConfigService } from '@/cli/utilities/config/services/config.service';
+import { ConfigService } from '@/cli/utilities/config/config-service';
 import axios, { type AxiosInstance, type AxiosResponse } from 'axios';
 import chalk from 'chalk';
 import * as fs from 'fs';
@@ -11,7 +11,7 @@ import {
 import * as path from 'path';
 import { type ApplicationManifest } from 'twenty-shared/application';
 import { type FileFolder } from 'twenty-shared/types';
-import { type ApiResponse } from '@/cli/utilities/api/types/api-response.types';
+import { type ApiResponse } from '@/cli/utilities/api/api-response-type';
 import { pascalCase } from 'twenty-shared/utils';
 
 export class ApiService {
