+++ packages/twenty-sdk/src/cli/utilities/config/current-execution-directory.ts
