+++ packages/twenty-sdk/src/cli/utilities/entity/__tests__/convert-to-label.spec.ts
@@ -1,4 +1,4 @@
-import { convertToLabel } from '@/cli/utilities/entity/utils/entity-label';
+import { convertToLabel } from '@/cli/utilities/entity/entity-label';
 
 describe('convertToLabel', () => {
   it('should convert to label', () => {
