+++ packages/twenty-sdk/src/cli/utilities/entity/__tests__/get-function-base-file.spec.ts
@@ -1,4 +1,4 @@
-import { getFunctionBaseFile } from '@/cli/utilities/entity/utils/entity-function-template';
+import { getFunctionBaseFile } from '@/cli/utilities/entity/entity-function-template';
 
 describe('getFunctionBaseFile', () => {
   it('should render proper file using defineFunction', () => {
