+++ packages/twenty-sdk/src/cli/utilities/entity/__tests__/get-role-base-file.spec.ts
@@ -1,4 +1,4 @@
-import { getRoleBaseFile } from '@/cli/utilities/entity/utils/entity-role-template';
+import { getRoleBaseFile } from '@/cli/utilities/entity/entity-role-template';
 
 describe('getRoleBaseFile', () => {
   it('should render proper file using defineRole', () => {
