+++ packages/twenty-sdk/src/cli/utilities/entity/entity-front-component-template.ts
