+++ packages/twenty-sdk/src/cli/utilities/entity/entity-function-template.ts
