+++ packages/twenty-sdk/src/cli/utilities/entity/entity-label.ts
