+++ packages/twenty-sdk/src/cli/utilities/entity/entity-object-template.ts
