+++ packages/twenty-sdk/src/cli/utilities/entity/entity-role-template.ts
