+++ packages/twenty-sdk/src/cli/utilities/typescript/typescript-diagnostics.ts
